+++ src/MintableCappedERC20.sol
@@ -2,10 +2,13 @@
 
 pragma solidity 0.8.9;
 
+import { IERC20BurnFrom } from './interfaces/IERC20BurnFrom.sol';
+
 import { ERC20 } from './ERC20.sol';
+import { ERC20Permit } from './ERC20Permit.sol';
 import { Ownable } from './Ownable.sol';
 
-contract MintableCappedERC20 is ERC20, Ownable {
+contract MintableCappedERC20 is ERC20, ERC20Permit, Ownable, IERC20BurnFrom {
     uint256 public cap;
 
     constructor(
@@ -13,7 +16,7 @@ contract MintableCappedERC20 is ERC20, Ownable {
         string memory symbol,
         uint8 decimals,
         uint256 capacity
-    ) ERC20(name, symbol, decimals) Ownable() {
+    ) ERC20(name, symbol, decimals) ERC20Permit(name) Ownable() {
         cap = capacity;
     }
 
@@ -22,5 +25,11 @@ contract MintableCappedERC20 is ERC20, Ownable {
         require(capacity == 0 || totalSupply + amount <= capacity, 'CAP_EXCEEDED');
 
         _mint(account, amount);
+    }
+
+    // TODO move burnFrom into a separate BurnableERC20 contract
+    function burnFrom(address account, uint256 amount) external onlyOwner {
+        _approve(account, owner, allowance[account][owner] - amount);
+        _burn(account, amount);
     }
 }
